// src/core/enums/item-type.ts
export enum ItemType {
	Artifact = 'Artifact',
	Consumable1st = 'Consumable (1st echelon)',
	Consumable2nd = 'Consumable (2nd echelon)',
	Consumable3rd = 'Consumable (3rd echelon)',
	Consumable4th = 'Consumable (4th echelon)',
	ImbuedArmor = 'Imbued Armor',
	ImbuedImplement = 'Imbued Implement',
	ImbuedWeapon = 'Imbued Weapon',
	LeveledArmor = 'Leveled Armor',
	LeveledImplement = 'Leveled Implement',
	LeveledWeapon = 'Leveled Weapon',
	Leveled = 'Leveled Item',
	Trinket1st = 'Trinket (1st echelon)',
	Trinket2nd = 'Trinket (2nd echelon)',
	Trinket3rd = 'Trinket (3rd echelon)',
	Trinket4th = 'Trinket (4th echelon)',
}